pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.0.1"
description = "A configuration dashboard using Discord OAuth login."
authors = ["vcokltfre <vcokltfre@gmail.com>"]
repository = "https://github.com/vcokltfre/dashboard"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
fastapi = "^0.66.0"
python-dotenv = "^0.18.0"
asyncpg = "^0.23.0"
starlette-discord = "^0.1.1"
taskipy = "^1.8.1"
aiofiles = "^0.7.0"
Jinja2 = "^3.0.1"
aioredis = "^1.3.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[tool.taskipy.tasks]
watch = "uvicorn src.api.main:app --reload"
start = "uvicorn src.api.main:app --host 0.0.0.0 --port 8087"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
